voiceinserter/requirements.txt
pyyaml
requests
joblib
pykakasi